+++ src/net/dinkla/raytracer/cameras/lenses/FishEye.java
@@ -14,22 +14,22 @@ import net.dinkla.raytracer.utilities.Resolution;
  */
 public class FishEye extends AbstractLens {
 
-    public float maxPsi;
+    public double maxPsi;
     
     public class RayDirection {
         public Vector3D direction = null;
-        public float rSquared = 0;
+        public double rSquared = 0;
     }
 
     public FishEye(ViewPlane viewPlane) {
         super(viewPlane);
-        maxPsi = 1.0f;
+        maxPsi = 1.0;
     }
 
     public Ray getRaySampled(int r, int c, Point2D sp) {
         Ray ray = null;
-        final float x = viewPlane.size * (c - 0.5f * viewPlane.resolution.hres + sp.getX());
-        final float y = viewPlane.size * (r - 0.5f * viewPlane.resolution.vres + sp.getY());
+        final double x = viewPlane.size * (c - 0.5 * viewPlane.resolution.hres + sp.getX());
+        final double y = viewPlane.size * (r - 0.5 * viewPlane.resolution.vres + sp.getY());
         final Point2D pp = new Point2D(x, y);
         RayDirection rd = getRayDirection(pp, viewPlane.resolution, viewPlane.size);
         if (rd.rSquared <= 1) {
@@ -40,8 +40,8 @@ public class FishEye extends AbstractLens {
 
     public Ray getRaySingle(int r, int c) {
         Ray ray = null;
-        final float x = viewPlane.size * (c - 0.5f * viewPlane.resolution.hres);
-        final float y = viewPlane.size * (r - 0.5f * viewPlane.resolution.vres);
+        final double x = viewPlane.size * (c - 0.5 * viewPlane.resolution.hres);
+        final double y = viewPlane.size * (r - 0.5 * viewPlane.resolution.vres);
         final Point2D pp = new Point2D(x, y);
         RayDirection rd = getRayDirection(pp, viewPlane.resolution, viewPlane.size);
         if (rd.rSquared <= 1) {
@@ -50,14 +50,14 @@ public class FishEye extends AbstractLens {
         return ray;
     }
 
-    protected RayDirection getRayDirection(Point2D pp, Resolution resolution, float s) {
+    protected RayDirection getRayDirection(Point2D pp, Resolution resolution, double s) {
         RayDirection rd = new RayDirection();
-        float x = 2.0f / (s * resolution.hres) * pp.getX();
-        float y = 2.0f / (s * resolution.vres) * pp.getY();
-        float rSquared = x * x + y * y;
+        double x = 2.0f / (s * resolution.hres) * pp.getX();
+        double y = 2.0f / (s * resolution.vres) * pp.getY();
+        double rSquared = x * x + y * y;
         if (rSquared <= 1) {
             float r = (float) Math.sqrt(rSquared);
-            float psi = r * maxPsi * MathUtils.PI_ON_180;
+            double psi = r * maxPsi * MathUtils.PI_ON_180;
             float sinPsi = (float) Math.sin(psi);
             float cosPsi = (float) Math.cos(psi);
             float sinAlpha = (float) y / r;
